docs(chore): Presentation - Wrap content with scope border
Stylistic choice, could be reverted. This was done to better visually scope the boundary of content tabs, as it's less obvious where they actually end when they're not within an admonition block.

This commit only adds the `!!! quote ""` lines with associated md comment above them, everything else is just indentation.