Update podman.md
Generate systemd is deprecated and Quadlets are now used. I have spent some time getting Quadlet to work, so I added some findings to the documentation. It is my first time contributing to documentation, so excuse my English if it comes across a bit rambly.

I have only added information, and have not removed any. However, generate-systemd is now deprecated and slirp4netns is out in favour of pasta. Pasta in particular is a lot easier to configure, but is still complicated to include into a docker-compose, which I do not use anymore.

Using podlet you can also translate a docker-compose to a Quadlet. Quadlets themselves behave a lot like individual  compose files, but are managed by systemd.